removed unusable credential check function
Kanbanize doesn't have a great way to validate that credentials are valid outside of it's login endpoint which just returns the apikey param anyway so it's probably not worth it to ask for username and password and instead just assume the apikey is correct and let the endpoint throw an error when it's queried and then handle that error appropriately. Whew, run on sentence.